// app/shared/model/conducteur.ts
import { EmployeSite } from './employe-site';

export class Conducteur {
    conducteurId: string;
    prenom: string;
    nom: string;
    image: string;
    cin: string;
    sexe: string;
    nationalite: string;
    date_Naisance: string;
    adresse: string;
    ville: string;
    province: string;
    codepostal: string;
    cellulaire: string;
    telephone: string;
    contact_Urgence: string;
    tel_Urgence: string;
    courriel: string;
    note: string;
    statut: string;
    userModification: string;
    dateModification: string;

   

   // employe_site: EmployeSite;
}